Adds MISC_SHARES_VIA_AGGREGATE standard trait
As discussed in Atlanta's PTG meetings, we will use the newly-added traits
functionality to mark a resource provider as providing shared resources to a
set of other resource providers. We need a new standard trait to represent this
shared provider characteristic, and this patch adds it
MISC_SHARES_VIA_AGGREGATE along with a fairly lengthy explanation of the
trait's meaning.

Change-Id: I4e9d8389c9b93ffee20adb53d864d55d335a27fd
blueprint: shared-resources-pike