Adds a validation test for trait names
Standard trait names should follow the regex "^[A-Z][A-Z0-9]*$". One of the
traits (HW_CPU_X86_AES-NI) erroneously contained a hyphen, so we correct that
and add a test to ensure no future traits wander from the standard trait name
regex.

Luckily, nothing yet is consuming os-traits and so the AES-NI slipup is yet to
have any impact.

Change-Id: I0603bcb4a1038289d2a3b6ce18aa823c7fac0829