Correctly recurse os_traits packages
When walk_packages is called, unless a prefix is provided, it is
easy to escape the os_traits package and start walking other
packages. If there is a top-level 'tests' package somewhere in the
python path, this can then lead to importing all sorts of things,
some which rely on the environment having certain characteristics
and fail the import without that, some which will change the
environment.

Neither of these things are good. Nor do we want to be walking
packages all over the place. By providing a prefix we stay within
the os_traits package as desired.

Change-Id: Icde41191e39317ff5a031fff9ece0f05f1e0b4bc
Closes-Bug: #1694792