Add a trait for the "bochs" display device
The 'bochs' display device is a recommended[1] safe option for UEFI
guests.

This lets an admin set the trait either via the flavor extra-specs or
image metadata properties:

    trait:COMPUTE_GRAPHICS_MODEL_BOCHS=required

This also allows The libvirt virt dirver report a standard trait
for bochs graphics model support and enables the image_metadata_prefilter
to automatically request the trait. This will enable vm that request
the bochs graphics model to be automaticaly scheduled to hosts that
can support it.

Implements: blueprint add-bochs-display-device
[1]: https://www.kraxel.org/blog/2019/09/display-devices-in-qemu/#tldr
Change-Id: Iba867f4632853f99373976e39c3494ea1c8b91a2
Signed-off-by: Kashyap Chamarthy <kchamart@redhat.com>